+++ paper/paper.bib
@@ -8,12 +8,12 @@
   }
   
 @Manual{ludecke2019see,
-title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
-author = {Daniel Lüdecke and Philip Waggoner and Mattan S. Ben-Shachar and Dominique Makowski},
-year = {2019},
-note = {R package version 0.2.0.9000},
-url = {https://easystats.github.io/see/},
-}
+    title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
+    author = {Daniel Lüdecke and Philip Waggoner and Mattan S. Ben-Shachar and Dominique Makowski},
+    year = {2019},
+    note = {R package version 0.2.0.9000},
+    url = {https://easystats.github.io/see/},
+    }
 
 @article{ludecke2019insight,
     journal = {Journal of Open Source Software},
@@ -71,7 +71,8 @@ url = {https://easystats.github.io/see/},
   number={1},
   pages={294--309},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.3758/s13428-016-0702-8}
 }
 
 @article{fieller1957tests,
@@ -82,7 +83,8 @@ url = {https://easystats.github.io/see/},
   number={3/4},
   pages={470--481},
   year={1957},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.1093/biomet/48.1-2.29}
 }
 
 @Manual{BayesFactor,
