### 2020-05-25 36a28a9
add formula for Kendall
