### 2020-05-25 39b0911
try latex command for smaller text
