+++ paper/paper.bib
@@ -75,6 +75,17 @@
   doi={10.3758/s13428-016-0702-8}
 }
 
+@article{langfelder2012fast,
+  title={Fast R functions for robust correlations and hierarchical clustering},
+  author={Langfelder, Peter and Horvath, Steve},
+  journal={Journal of statistical software},
+  volume={46},
+  number={11},
+  year={2012},
+  publisher={NIH Public Access}
+}
+
+
 @article{fieller1957tests,
   title={Tests for rank correlation coefficients. I},
   author={Fieller, Edgar C and Hartley, Herman O and Pearson, Egon S},
