### 2020-05-25 027d3f3
smaller code output
