### 2020-06-07 #2233 @markhwhiteii
Hi all,

Glad to help review here, but this would be my first time doing a review for JOSS. Is there a specific way this journal does peer-review? What is the typical procedure?
