### 2020-06-07 #2233 @arfon
> Hi all,

> Glad to help review here, but this would be my first time doing a review for JOSS. Is there a specific way this journal does peer-review? What is the typical procedure?

:wave: @markhwhiteii - thanks for being willing to edit this. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .

Before the review will start though, the editor (@mikldk) will need to find a second reviewer for this submission.
