### 2020-06-08 #2306 @whedon
[REVIEW]: Methods and Algorithms for Correlation Analysis in R
**Submitting author:** @DominiqueMakowski (<a href="http://orcid.org/0000-0001-5375-9967">Dominique Makowski</a>)
**Repository:** <a href="https://github.com/easystats/correlation" target ="_blank">https://github.com/easystats/correlation</a>
**Version:** v0.3.0
