### 2020-06-08 #2306 @mikldk
@DominiqueMakowski: Thank you for the submission. I have now started the review.

@markhwhiteii, @mmrabe: Thanks for agreeing to review. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.

@DominiqueMakowski, @markhwhiteii, @mmrabe: The paper seems fairly long. It could be an idea to make it shorter (maybe by moving content to the vignettes). The reviewers might have more specific comments on this so I will let them have a look at it.

If any of you have any questions or concerns please let me know.
