### 2020-06-18 #2306 @markhwhiteii
- There's no reference to or explanation of Figure 1 in the text.
    It is a little confusing, as it is a barplot with a scatterplot
    layered over it. Given the length of the document, I would
    suggest perhaps removing Figure 1.
- The print output of calls to `correlation()` run off the edge of
    the page. Is there a way to shrink the text size or to cut off
    the columns that run wider than the rest of the text?
- The doi for Langfelder & Horvath (2012) seems to be missing, but
    I did some Googling, and it does appear as though there is not
    one.
- The second figure, the correlation heatmap, does not have a
    figure number or caption corresponding to it.
- The third figure is listed as the second figure, and it has a
    redundant figure number in the caption.
- I think it would be useful to readers to include a citation to a
    high-level explanation of GGM in the GGM section.
- Similarly, I think it would be useful to provide a citation in
    the multilevel model section so that the reader can look into
    the details of how these correlations are calculated, if they
    would like to know.
