### 2020-06-18 easystats/correlation#85 @DominiqueMakowski
JOSS review: comment 1
https://github.com/openjournals/joss-reviews/issues/2306#issuecomment-645712139

> There's no reference to or explanation of Figure 1 in the text.
It is a little confusing, as it is a barplot with a scatterplot
layered over it. Given the length of the document, I would
suggest perhaps removing Figure 1.
