### 2020-06-18 easystats/correlation#86 @DominiqueMakowski
JOSS review: comment 2
openjournals/joss-reviews#2306 (comment)

> The print output of calls to correlation() run off the edge of
the page. Is there a way to shrink the text size or to cut off
the columns that run wider than the rest of the text?
