### 2020-06-18 easystats/correlation#87 @DominiqueMakowski
JOSS review: comment 3
openjournals/joss-reviews#2306 (comment)

> The doi for Langfelder & Horvath (2012) seems to be missing, but
I did some Googling, and it does appear as though there is not
one.
