### 2020-06-18 easystats/correlation#88 @DominiqueMakowski
JOSS review: comment 4
openjournals/joss-reviews#2306 (comment)

> The second figure, the correlation heatmap, does not have a
figure number or caption corresponding to it.
