### 2020-06-18 easystats/correlation#89 @DominiqueMakowski
JOSS review: comment 5
openjournals/joss-reviews#2306 (comment)

> The third figure is listed as the second figure, and it has a
redundant figure number in the caption.
