### 2020-06-18 easystats/correlation#90 @DominiqueMakowski
JOSS review: comment 6
openjournals/joss-reviews#2306 (comment)

> I think it would be useful to readers to include a citation to a
high-level explanation of GGM in the GGM section.
