### 2020-06-18 easystats/correlation#91 @DominiqueMakowski
JOSS review: comment 7
openjournals/joss-reviews#2306 (comment)

> Similarly, I think it would be useful to provide a citation in
the multilevel model section so that the reader can look into
the details of how these correlations are calculated, if they
would like to know.
