### 2020-06-18 easystats/correlation#86 @strengejacke
`correlation(iris)[, c(1:5, 8)]`
