### 2020-06-22 c280bb9
improve fig1 legend and call it from text
#85
