### 2020-06-22 easystats/correlation#85 @DominiqueMakowski
We have improved the legend of figure 1 and also introduced it from the body. I think this figure serves as a nice visual "summary" and an illustration of the first section, which core point is that there are different measures that can give different results for the same data. Additionally, this figure can be used as a visual abstract of the article.
