+++ paper/paper.bib
@@ -82,7 +82,8 @@
   volume={46},
   number={11},
   year={2012},
-  publisher={NIH Public Access}
+  publisher={NIH Public Access},
+  url={http://www.jstatsoft.org/v46/i11/}
 }
 
 
