### 2020-06-22 easystats/correlation#87 @DominiqueMakowski
That is correct, the DOI does not seem to exist for this reference:

- https://pubmed.ncbi.nlm.nih.gov/23050260/
- https://cran.r-project.org/web/packages/flashClust/citation.html
