### 2020-06-22 5c07111
Added references for GGM #90
