+++ paper/paper.bib
@@ -32,6 +32,18 @@
     day = {25},
 }
 
+@article{Bhushan_2019,
+	doi = {10.3389/fpsyg.2019.01050},
+	url = {https://doi.org/10.3389%2Ffpsyg.2019.01050},
+	year = 2019,
+	month = {may},
+	publisher = {Frontiers Media {SA}},
+	volume = {10},
+	author = {Nitin Bhushan and Florian Mohnert and Daniel Sloot and Lise Jans and Casper Albers and Linda Steg},
+	title = {Using a Gaussian Graphical Model to Explore Relationships Between Items and Variables in Environmental Psychology Research},
+	journal = {Frontiers in Psychology}
+}
+
 
 @article{makowski2019bayestestr,
     title = {{bayestestR}: {Describing} {Effects} and their {Uncertainty}, {Existence} and {Significance} within the {Bayesian} {Framework}},
@@ -61,6 +73,20 @@
 	author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy McGowan and Romain Fran{\c{c}}ois and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Pedersen and Evan Miller and Stephan Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
 	title = {Welcome to the Tidyverse},
 	journal = {Journal of Open Source Software}
+}
+
+@article{epskamp2018estimating,
+	doi = {10.3758/s13428-017-0862-1},
+	url = {https://doi.org/10.3758%2Fs13428-017-0862-1},
+	year = 2018,
+	month = {feb},
+	publisher = {Springer Science and Business Media {LLC}},
+	volume = {50},
+	number = {1},
+	pages = {195--212},
+	author = {Sacha Epskamp and Denny Borsboom and Eiko I. Fried},
+	title = {Estimating psychological networks and their accuracy: A tutorial paper},
+	journal = {Behavior Research Methods}
 }
 
 @article{bishara2017confidence,
