### 2020-06-22 easystats/correlation#90 @DominiqueMakowski
We added a sentence and references to introductions to GGMs.
