### 2020-06-22 easystats/correlation#90 @strengejacke
Maybe we can add this somewhere to the package as well?
