### 2020-06-22 0a880a8
add GGM clarification to the docs and package README #90
