### 2020-06-22 easystats/correlation#88 @DominiqueMakowski
That has been resolved.
