### 2020-06-22 91357a3
try smaller size for tables #86
