+++ paper/paper.md
@@ -88,14 +88,14 @@ It relies on one main function, `correlation()`, which outputs a dataframe conta
 ``` r
 cor <- correlation(iris)
 cor
-## Parameter1   |   Parameter2 |     r |         95% CI |     t |  df |      p |  Method | n_Obs
-## ---------------------------------------------------------------------------------------------
-## Sepal.Length |  Sepal.Width | -0.12 | [-0.27,  0.04] | -1.44 | 148 | 0.152  | Pearson |   150
-## Sepal.Length | Petal.Length |  0.87 | [ 0.83,  0.91] | 21.65 | 148 | < .001 | Pearson |   150
-## Sepal.Length |  Petal.Width |  0.82 | [ 0.76,  0.86] | 17.30 | 148 | < .001 | Pearson |   150
-## Sepal.Width  | Petal.Length | -0.43 | [-0.55, -0.29] | -5.77 | 148 | < .001 | Pearson |   150
-## Sepal.Width  |  Petal.Width | -0.37 | [-0.50, -0.22] | -4.79 | 148 | < .001 | Pearson |   150
-## Petal.Length |  Petal.Width |  0.96 | [ 0.95,  0.97] | 43.39 | 148 | < .001 | Pearson |   150
+> Parameter1   |   Parameter2 |     r |         95% CI |     t |  df |      p |  Method | n_Obs
+> ---------------------------------------------------------------------------------------------
+> Sepal.Length |  Sepal.Width | -0.12 | [-0.27,  0.04] | -1.44 | 148 | 0.152  | Pearson |   150
+> Sepal.Length | Petal.Length |  0.87 | [ 0.83,  0.91] | 21.65 | 148 | < .001 | Pearson |   150
+> Sepal.Length |  Petal.Width |  0.82 | [ 0.76,  0.86] | 17.30 | 148 | < .001 | Pearson |   150
+> Sepal.Width  | Petal.Length | -0.43 | [-0.55, -0.29] | -5.77 | 148 | < .001 | Pearson |   150
+> Sepal.Width  |  Petal.Width | -0.37 | [-0.50, -0.22] | -4.79 | 148 | < .001 | Pearson |   150
+> Petal.Length |  Petal.Width |  0.96 | [ 0.95,  0.97] | 43.39 | 148 | < .001 | Pearson |   150
 ```
 
 \normalsize
@@ -106,11 +106,11 @@ The output is not a square matrix, but a **(tidy) dataframe with all correlation
 
 ``` r
 summary(cor)
-## Parameter    | Petal.Width | Petal.Length | Sepal.Width
-## -------------------------------------------------------
-## Sepal.Length |     0.82*** |      0.87*** |       -0.12
-## Sepal.Width  |    -0.37*** |     -0.43*** |            
-## Petal.Length |     0.96*** |              |
+> Parameter    | Petal.Width | Petal.Length | Sepal.Width
+> -------------------------------------------------------
+> Sepal.Length |     0.82*** |      0.87*** |       -0.12
+> Sepal.Width  |    -0.37*** |     -0.43*** |            
+> Petal.Length |     0.96*** |              |
 ```
 
 \normalsize
@@ -121,12 +121,12 @@ Note that one can also obtain the full, **square** and redundant matrix using:
 
 ``` r
 summary(cor, redundant=TRUE)
-## Parameter    | Sepal.Length | Sepal.Width | Petal.Length | Petal.Width
-## ----------------------------------------------------------------------
-## Sepal.Length |      1.00*** |       -0.12 |      0.87*** |     0.82***
[…]
