### 2020-06-22 easystats/correlation#89 @DominiqueMakowski
This has been fixed.
