### 2020-06-22 easystats/correlation#86 @DominiqueMakowski
This has been [fixed](https://github.com/openjournals/joss-papers/blob/joss.02306/joss.02306/10.21105.joss.02306.pdf), although for one table it's quite close the margin, but it' either that or we remove so column (and they all seem important in this case). @strengejacke what do you think?
