### 2020-06-22 easystats/correlation#86 @strengejacke
It's up to the journal editor to decide about the margins, but if we need to make the table smaller (narrower), I'd suggest removing `n_obs`, and if not enough, also "df". You could add a footnote saying that due to layout 1/2 columns were removed.
