### 2020-06-22 easystats/correlation#86 @DominiqueMakowski
do you suggest we leave it as is for now, and then if the editor finds it to be an issue we proceed and remove one or two cols?
