### 2020-06-22 easystats/correlation#86 @strengejacke
yes, I would say so.
