### 2020-06-22 easystats/correlation#86 @DominiqueMakowski
Sounds good to me
