### 2020-06-22 easystats/correlation#90 @DominiqueMakowski
This has been [addressed](https://github.com/openjournals/joss-papers/blob/joss.02306/joss.02306/10.21105.joss.02306.pdf).
