+++ paper/paper.bib
@@ -7,6 +7,19 @@
     url = {https://www.R-project.org/},
   }
   
+@article{shan2020correlation,
+	doi = {10.1155/2020/7398324},
+	url = {https://doi.org/10.1155%2F2020%2F7398324},
+	year = 2020,
+	month = {mar},
+	publisher = {Hindawi Limited},
+	volume = {2020},
+	pages = {1--11},
+	author = {Guogen Shan and Hua Zhang and Tao Jiang},
+	title = {Correlation Coefficients for a Study with Repeated Measures},
+	journal = {Computational and Mathematical Methods in Medicine}
+}
+
 @Manual{ludecke2019see,
     title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
     author = {Daniel Lüdecke and Philip Waggoner and Mattan S. Ben-Shachar and Dominique Makowski},
@@ -87,6 +100,18 @@
 	author = {Sacha Epskamp and Denny Borsboom and Eiko I. Fried},
 	title = {Estimating psychological networks and their accuracy: A tutorial paper},
 	journal = {Behavior Research Methods}
+}
+
+@article{bakdash2017repeated,
+	doi = {10.3389/fpsyg.2017.00456},
+	url = {https://doi.org/10.3389%2Ffpsyg.2017.00456},
+	year = 2017,
+	month = {apr},
+	publisher = {Frontiers Media {SA}},
+	volume = {8},
+	author = {Jonathan Z. Bakdash and Laura R. Marusich},
+	title = {Repeated Measures Correlation},
+	journal = {Frontiers in Psychology}
 }
 
 @article{bishara2017confidence,
