### 2020-06-22 easystats/correlation#91 @DominiqueMakowski
We added some useful references accordingly.
