### 2020-06-29 #2306 @DominiqueMakowski
Dear @markhwhiteii, thanks for your comments and suggestions ☺️ We have addressed and fixed all of your points in separate issues (see https://github.com/easystats/correlation/issues?q=is%3Aissue+is%3Aclosed). Let us know if there is anything more.
