### 2020-06-30 #2306 @mmrabe
I was now able to complete my review. All in all, the paper is a very useful introduction to the package. I don't have a lot to add to what @markhwhiteii already mentioned and has been fixed in the meantime. I only have a few minor suggestions, which I added as Github issues. I do agree with @mikldk that the paper is a bit too long for a JOSS submission. I would recommend selecting two representative examples and maybe shorten the very informative but too detailed intros to different types of correlations.
