### 2020-06-30 a48c637
cleanup Spearman's rank correlation
