### 2020-06-30 a009f38
Add Vignettes to package (I have no idea how come it wans't there) #95
