+++ DESCRIPTION
@@ -38,6 +38,7 @@ Suggests:
 	BayesFactor,
 	dplyr,
 	energy,
+	forcats,
 	ggcorrplot,
 	ggplot2,
 	Hmisc,
@@ -57,4 +58,5 @@ Suggests:
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
+VignetteBuilder: knitr
