### 2020-06-30 a09a712
add warning for when not enough variables
close #92
