### 2020-06-30 74b6411
'%' -> 'percent' #93
