### 2020-06-30 9bcdd81
replaced 'than' #93
