### 2020-06-30 72c91a8
replace 'cor' by 'results' #94
