+++ man/correlation.Rd
@@ -99,11 +99,11 @@ Such partial correlations can be represented as Gaussian Graphical Models (GGM),
 }
 \examples{
 library(correlation)
-cor <- correlation(iris)
+results <- correlation(iris)
 
-cor
-summary(cor)
-summary(cor, redundant = TRUE)
+results
+summary(results)
+summary(results, redundant = TRUE)
 
 # Grouped dataframe
 if (require("dplyr")) {
