### 2020-07-01 easystats/correlation#96 @DominiqueMakowski
Shorten paper
As mentioned by both reviewers (https://github.com/openjournals/joss-reviews/issues/2306), shorten the paper.
