### 2020-07-01 easystats/correlation#96 @DominiqueMakowski
@mmrabe following your suggestions, we tried to see how to address best and here are some thoughts:

- **Shorten the output tables**: some of them are indeed lengthy (in particular, one), but all of them show something that justifies their length. For instance, the longest table shows (https://github.com/openjournals/joss-papers/blob/joss.02306/joss.02306/10.21105.joss.02306.pdf) showcases the "auto" method, and thus if we cut the table it doesn't show the automatic selection of the different methods. As for the other tables, they are either square matrics or short tables, which cutting doesn't really make sense.
- **Remove some examples**: As they are already short and really "illustrative", removing any feels a bit arbitrary to me.
- **Remove some figures**: While we could remove some of the plotting examples, such visuals are important for readers and they do not (IMO) "crowd" the paper.
- **Remove the "correlation types" description**: this is one of the important features of the package, and removing this (already very brief) description seems a bit counterintuitive. Moreover, because of their bullet list presentation, the reader can just read whatever point he wants, he doesn't have to read entire paragraphs to gain relevant insights.

I would suggest leaving the decision to the editor and see if he's okay with the current paper length.
