+++ NEWS.md
@@ -1,6 +1,7 @@
 # correlation 0.x.x
 
 ## Changes
+- `adjust()` has been moved over from `effectsize` 
 
 ## Bug fixes
 

[…]
