### 2020-07-06 #2306 @mikldk
@DominiqueMakowski @mmrabe @markhwhiteii : I am worried with the length of the paper. It is currently approx 2,500 words and <https://joss.readthedocs.io/en/latest/submitting.html> says "The paper should be between 250-1000 words.". The [issue](https://github.com/easystats/correlation/issues/96) did not track commits so it is difficult to follow what changes was made. I suggest that author reopens the issue and work on moving content from paper to vignettes and also re-read <https://joss.readthedocs.io/en/latest/submitting.html>. What do you think @mmrabe and @markhwhiteii?
