### 2020-07-07 #2306 @DominiqueMakowski
Thanks @mikldk for your feedback. We'll try shortening it. However, note that the current length and format of the paper was mostly inspired by the previous JOSS papers of the other easystats papers (https://joss.theoj.org/papers/10.21105/joss.01541; https://joss.theoj.org/papers/10.21105/joss.01412) which length also exceeds the said recommendation. Our concern is that the paper is typically aimed at a different audience (for instance posted on researchgate), such as academics, that are less used to explore the features on GitHub and in the vignettes. Hence we wanted the paper to also showcase some of the features (rather than be just a container of links), but we'll see what we can do ☺️
