### 2020-07-07 #2306 @mmrabe
@mikldk Are there more detailed guidelines what exactly counts toward the 250-1000 words or how one should determine the word count? Ignoring figures, output, and equations, I count approx. 1080 words.

@DominiqueMakowski I've also noticed that many R users won't look at vignettes but given that your audience does care about papers, wouldn't the paper be a good place to make them aware of that feature? You should also consider that your paper will remain static but you can always add, extend, and revise package vignettes after publication. That may be advantageous, too.
