### 2020-07-08 #2306 @mikldk
@mmrabe: I simply took the markdown without the header.

@DominiqueMakowski : Please put examples in the software documentation and link to that in the paper. This can both be in vignettes or in the README. Also only include high-level summary of the different functionalities that the package supports in the paper and refer to the documentation for a complete summary of functionality.
