### 2020-07-09 87634b8
shorten paper #96
