+++ paper/paper.md
@@ -79,9 +79,7 @@ These methods allow for different ways of quantifying the link between two varia
 
 It relies on one main function, `correlation()`, which outputs a dataframe containing each pairwise correlation per row. This long format is convenient for further data analysis, but not as much to get a summary, which is usually obtained via a correlation matrix. To address this, we added standard methods, such as `summary()` and `as.matrix()`, to automatically transform the long output to a matrix. Moreover, **correlation** also includes plotting capabilities via the [**see** package](https://easystats.github.io/see/) [@ludecke2019see].
 
-# Examples
-
-## Correlation details and matrix
+An overview of the features is available on the GitHub page (https://github.com/easystats/correlation). The typical core workflow is as follows:
 
 \footnotesize
 
@@ -100,7 +98,7 @@ results
 
 \normalsize
 
-The output is not a square matrix, but a **(tidy) dataframe with all correlations tests per row**. One can also obtain a **matrix** using:
+The output is not a square matrix, but a (tidy) dataframe with all correlations tests per row. One can also obtain a matrix using:
 
 \small
 
@@ -115,191 +113,9 @@ summary(results)
 
 \normalsize
 
-Note that one can also obtain the full, **square** and redundant matrix using:
-
-\small
-
-``` r
-summary(results, redundant=TRUE)
-# Parameter    | Sepal.Length | Sepal.Width | Petal.Length | Petal.Width
-# ----------------------------------------------------------------------
-# Sepal.Length |      1.00*** |       -0.12 |      0.87*** |     0.82***
-# Sepal.Width  |        -0.12 |     1.00*** |     -0.43*** |    -0.37***
-# Petal.Length |      0.87*** |    -0.43*** |      1.00*** |     0.96***
-# Petal.Width  |      0.82*** |    -0.37*** |      0.96*** |     1.00***
-```
-
-``` r
-library(dplyr)
-library(see)
-
-results %>% 
-  summary(redundant=TRUE) %>% 
-  plot()
-```
-
-\normalsize
-
-![Correlation matrix plot that can be automatically obtained via the **see** package.](figure2.png)
 
 
-## Grouped dataframes
-
-The function also supports **stratified correlations**, all within the *tidyverse* [@Wickham_2019] workflow.
-
-\scriptsize
-
-``` r
-iris %>% 
-  select(Species, Sepal.Length, Sepal.Width, Petal.Width) %>% 
-  group_by(Species) %>% 
-  correlation()
-# Group      |   Parameter1 |  Parameter2 |    r |        95% CI |    t | df |      p |  Method | n_Obs
-# -----------------------------------------------------------------------------------------------------
-# setosa     | Sepal.Length | Sepal.Width | 0.74 | [ 0.59, 0.85] | 7.68 | 48 | < .001 | Pearson |    50
-# setosa     | Sepal.Length | Petal.Width | 0.28 | [ 0.00, 0.52] | 2.01 | 48 | 0.101  | Pearson |    50
-# setosa     |  Sepal.Width | Petal.Width | 0.23 | [-0.05, 0.48] | 1.66 | 48 | 0.104  | Pearson |    50
-# versicolor | Sepal.Length | Sepal.Width | 0.53 | [ 0.29, 0.70] | 4.28 | 48 | < .001 | Pearson |    50
-# versicolor | Sepal.Length | Petal.Width | 0.55 | [ 0.32, 0.72] | 4.52 | 48 | < .001 | Pearson |    50
-# versicolor |  Sepal.Width | Petal.Width | 0.66 | [ 0.47, 0.80] | 6.15 | 48 | < .001 | Pearson |    50
[…]
