### 2020-07-09 #2306 @DominiqueMakowski
Thanks for your suggestions.

@mikldk we shorten the paper and linked to the vignettes and documentation.
