### 2020-07-09 #2306 @mikldk
@DominiqueMakowski : Thank you for the progress. The vignette "Correlation Types" does not contain as many details (formulas) as the paper. I think it should be the other way around. Also, you still have `R` examples in the paper. Please see if you can improve the vignette(s) with what is in the paper (e.g. details on correlation) and try to keep the paper without examples (move to vignettes/README).
