### 2020-07-10 e0a84a5
more details about types in vignettes
