+++ vignettes/bibliography.bib
@@ -1,0 +1,159 @@
+@Manual{Rteam,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2019},
+    url = {https://www.R-project.org/},
+  }
+  
+@article{shan2020correlation,
+	doi = {10.1155/2020/7398324},
+	url = {https://doi.org/10.1155%2F2020%2F7398324},
+	year = 2020,
+	month = {mar},
+	publisher = {Hindawi Limited},
+	volume = {2020},
+	pages = {1--11},
+	author = {Guogen Shan and Hua Zhang and Tao Jiang},
+	title = {Correlation Coefficients for a Study with Repeated Measures},
+	journal = {Computational and Mathematical Methods in Medicine}
+}
+
+@Manual{ludecke2019see,
+    title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
+    author = {Daniel Lüdecke and Philip Waggoner and Mattan S. Ben-Shachar and Dominique Makowski},
+    year = {2019},
+    note = {R package version 0.2.0.9000},
+    url = {https://easystats.github.io/see/},
+    }
+
+@article{ludecke2019insight,
+    journal = {Journal of Open Source Software},
+    doi = {10.21105/joss.01412},
+    issn = {2475-9066},
+    number = {38},
+    publisher = {The Open Journal},
+    title = {insight: A Unified Interface to Access Information from Model Objects in R},
+    url = {http://dx.doi.org/10.21105/joss.01412},
+    volume = {4},
+    author = {Lüdecke, Daniel and Waggoner, Philip and Makowski, Dominique},
+    pages = {1412},
+    date = {2019-06-25},
+    year = {2019},
+    month = {6},
+    day = {25},
+}
+
+@article{Bhushan_2019,
+	doi = {10.3389/fpsyg.2019.01050},
+	url = {https://doi.org/10.3389%2Ffpsyg.2019.01050},
+	year = 2019,
+	month = {may},
+	publisher = {Frontiers Media {SA}},
+	volume = {10},
+	author = {Nitin Bhushan and Florian Mohnert and Daniel Sloot and Lise Jans and Casper Albers and Linda Steg},
+	title = {Using a Gaussian Graphical Model to Explore Relationships Between Items and Variables in Environmental Psychology Research},
+	journal = {Frontiers in Psychology}
+}
+
+
+@article{makowski2019bayestestr,
+    title = {{bayestestR}: {Describing} {Effects} and their {Uncertainty}, {Existence} and {Significance} within the {Bayesian} {Framework}},
+    volume = {4},
+    issn = {2475-9066},
+    shorttitle = {{bayestestR}},
+    url = {https://joss.theoj.org/papers/10.21105/joss.01541},
+    doi = {10.21105/joss.01541},
+    number = {40},
+    urldate = {2019-08-13},
+    journal = {Journal of Open Source Software},
+    author = {Makowski, Dominique and Ben-Shachar, Mattan and Lüdecke, Daniel},
+    month = aug,
+    year = {2019},
+    pages = {1541}
+}
+
+@article{Wickham_2019,
+	doi = {10.21105/joss.01686},
+	url = {https://doi.org/10.21105%2Fjoss.01686},
+	year = 2019,
+	month = {nov},
+	publisher = {The Open Journal},
+	volume = {4},
+	number = {43},
+	pages = {1686},
[…]
