### 2020-07-10 #2306 @DominiqueMakowski
Thanks for your feedback @mikldk. We have addressed the issues regarding the vignettes so that they include all the details about the types.

Regarding the two `R` chunks that we left in the paper, I would like to ask for the reconsideration of the possibility of keeping them, as they are not really "examples" of features per se but rather just the demonstration of the main function and its output. I think it's important that the audience reading the paper can have a glimpse of how the workflow looks like, so that they can decide if they want to further explore the features by visiting the documentation/vignettes.
