### 2020-07-10 #2306 @mmrabe
I agree with @DominiqueMakowski that it would probably be helpful for most readers if the paper contains a very brief demonstration of the main functionality, as it does now.
