### 2020-07-14 #2306 @mikldk
@DominiqueMakowski :

* Please add DOI 10.18637/jss.v046.i11 for `@langfelder2012fast`
* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
