### 2020-07-14 #2306 @DominiqueMakowski
Dear @mikldk, regarding the first point, as mentioned [here](https://github.com/easystats/correlation/issues/87), we weren't able to find a DOI for this ref.

We checked the paper and the metadata.

The zenodo release version is 0.3.0 and its number: **10.5281/zenodo.3943771** (https://zenodo.org/record/3943771).

Thanks a lot.
