### 2020-07-14 #2306 @mmrabe
@DominiqueMakowski
The DOI `10.18637/jss.v046.i11` that @mikldk mentioned is listed under the URL linked in the [package citation](https://cran.r-project.org/web/packages/flashClust/citation.html). Also, while looking at the bibliography again, could you maybe escape all occurences of {R} in your .bib so that it appears in upper case in the rendered version?
