### 2020-07-14 764663b
fix R in uppercase
https://github.com/openjournals/joss-reviews/issues/2306
