+++ paper/paper.bib
@@ -127,7 +127,7 @@
 }
 
 @article{langfelder2012fast,
-  title={Fast R functions for robust correlations and hierarchical clustering},
+  title={Fast {R} functions for robust correlations and hierarchical clustering},
   author={Langfelder, Peter and Horvath, Steve},
   journal={Journal of statistical software},
   volume={46},
