### 2020-07-14 #2306 @DominiqueMakowski
> The DOI 10.18637/jss.v046.i11 that @mikldk mentioned

Apologies, I misread this and missed the DOI given by @mikldk ☺️

> escape all occurences of {R} in your .bib

It should be done, thansk a lot @mmrabe !!
