### 2020-07-14 996ef3b
add missing DOI to ref
https://github.com/openjournals/joss-reviews/issues/2306
