+++ paper/paper.bib
@@ -134,7 +134,8 @@
   number={11},
   year={2012},
   publisher={NIH Public Access},
-  url={http://www.jstatsoft.org/v46/i11/}
+  url={http://www.jstatsoft.org/v46/i11/},
+  doi={10.18637/jss.v046.i11}
 }
 
 
