### 2020-07-14 #2306 @DominiqueMakowski
@mikldk please note that the version is 0.3.0 (I think you entered 0.0.3)
